requirements.txt
aniso8601==9.0.1
autopep8==1.5.7
backcall==0.2.0
click==8.0.1
decorator==5.0.9
greenlet==1.1.0
ipdb==0.13.7
ipython==7.23.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.25
mysqlclient==1.3.13
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
protobuf==3.17.0
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
